fix($routing): Fix endpoint routing w/ params
When performing an HTTP request with the incorrect
method (verb) to an endpoint the expected result should
be an HTTP 405. However the method_not_allowed() method
that is dynamically assigned to each Resource object
does not have a variable-length argument parameter to
handle these objects. This pull request simply adds (but
ignores) the parameter and provides a corresponding unit
test.