fix(Request): Made header and query param parsing more strict
This patch updates several helper properties and methods on the Request
object to raise HTTPBadRequest when the value of the field is not
in the expected format. If it should be an int, then it must be an int. If
Content-Range is not in a supported format, the request is denied.

BREAKING CHANGE: Whereas previously parsing an incorrectly-formatted header
or query parameter would result in a None response, attempting to parse
such values now raises HTTPBadRequest instead.

Fixes #99