refactor(bench): Moved folder under falcon/ and added entrypoint to setup.py
This patch moves the benchmark folder down a level (under falcon/), and
cleans it up a bit. There is now an entrypoint that gets installed, called
falcon-bench, to make it easier to run benchmarks. Falcon-bench will print
out a message for every framework it can't import; by default only Falcon
can be benchmarked.