feat(api): add_sink method
This patch adds a new method, add_sink(...), to the falcon.API class.

EXAMPLES
========

Any requested URI path requested that starts with '/my-proxy' will be
passed to my_proxy, regardless of HTTP verb.

    app.add_sink(my_proxy, prefix='/my-proxy')

Any URI path requested that doesn't match a route will be
passed to my_proxy, regardless of HTTP verb.

    app.add_sink(my_proxy)

In the case of a collision between a route and a sync, the route
always wins.

    app.add_route('/books', book_collection)
    app.add_sink(book_worm, '/books')

The prefix may either be a regex string or a precompiled regex object:

    app.add_sink(my_proxy, prefix=r'/my-proxy/\d+')
    app.add_sink(my_thingy, prefix=re.compile('/thingy/[a-zA-Z]+'))