Fix inaccurate docstring for API._get_responder
The docstring for falcon.api.API._get_responder incorrectly stated
that the return from the function was a 3-tuple. This patch corrects
that to state that the return is a 2-tuple containing the responder
and the param dict, and adds some notes about the responder returned
for bad requests and no matching route.