doc(reference): Improved and annotated Request docstrings
The Request docstrings were modified so that now attributes and properties
are both documented in the class docstring, and class members were
reordered to be grouped more logically so that browsing the source as
well as reading the docs is easier.

Also, None was annotated across multiple files, and some minor tweaks to
introductory prose were completed.