feat(API): Support custom Request and Response classes
This patch adds support for supplying custom request and
response classes.

As part of this work, a new "context" attribute was added
to Request, and its type may be overridden by setting the
"context_type" class attribute in a custom Request class.

The alternative to using a class attribute would have been
to pass the context type into the api, and then each time
a class was instantiated, pass the type into it's
initializer. After discussion on IRC, it was decided that
using a class attribute would be more elegant.

See commentary at: https://github.com/racker/falcon/pull/256

Closes #256 and #248