feat(hooks): Pass resource to hooks
Make "before" and "after" hooks aware of the resource being acted
upon. This was implemented in such a way that existing hooks will
still work; the framework will detect whether a given hook function
accepts the extra "resource" param, and wire it up accordingly.

Rationale: This change is intended to ease creating more descriptive
and declarative resources. Hooks are the recommended approach for
implementing creating authentication/authorization solutions. However,
Authorization almost always depends on which resource is being
accessed. This patch provides the necessary context for such use
cases.

A good example would be an OAuth (1.0/2.0) where there is an concept
of "scopes" or "realms". Good OAuth middleware should allow
declarations like this:

class Resource(object):
    realms_allowed = ['admin', 'editor', 'viewer']

    @realms('viewer')
    def on_get(req, resp):
        pass

With this patch it will be easier to implement sophisticated
middleware.

Closes #268