These unicode declarations prevent importing of falcon, under python 3.2.
This is a quick and dirty work-around, and obviously sub-optimal. The best option, is if there's a way to rewrite the code in request.py so it passes syntax validation in both versions of python. Perhaps instead of using u to declare it unicode, you do it in ascii, and convert it to unicode if it's python2x? It doesn't look like there's any unicode requiring characters in there.