feat(hooks): Run global "after" hooks even when an exception is raised
This patch causes Falcon to execute the list of global "after" hooks
after an exception is raised. Before this change, exceptions would
circumvent the hooks, so there was no opportunity to install, for
example, a custom serializer that could also act on errors.

Note that per-resource and per-responder "after" hooks will continue to
be circumvented after applying this patch; it was deemed to hacky
to make those work. If an elegant solution can be found, and there is
a good use case for doing it, we can revisit this issue.

Closes #250