add a new API which allows one class to implement many web methods
current API only allows one class to implement web methods according
to its method names, this is limitting in some cases. With this new
API, a developer can use one class to implement a set of APIs.
The developer can simply add an instance of that class to the app
by calling add_route method. An example of how to do this has been
documented in the ResourceAPI class in resource_api.py module.