current API only allows a class to implement web methods by manipulating the method name, such as on_get, on_post etc. This approach is rather limitting in some cases. With this new extended API, a developer can use one class to implement many methods. This is particularly important for an application which wants to support multiple implementations of a set of APIs.
With this new class, a developer can simply add an instance of
resouce class which may implement all the API with the API
required url patterns, then use that class to control all its
operations.

An example of how to use this feature has been documented
in the ResourceAPI class in resource_api.py module.