bugfix(errors): re-enable body on the messages
Change the 404 and 405 errors to allow an optional representation.
Developers should be able to decide whether or not to send the error
message body.  OptionalRepresentation by default behaves like NoRepresentation.
If all were switched, then NoRepresentation should be left for devs to use
in their own error derived classes to force it one way all the time.

No breaking changes