feat(errors): use super()
Updated HTTP 404 and 405 errors to use super() instead of directly calling
the initializer for other objects. This allows Pytons MRU functionality to
do the heavy lifting and ensure this work correctly. It also enables us to
not have to initialize self.description in the OptionalRepresentation mix-in
as it can be retrieved from other classes where it does get initialized ;-)