feat(routing): Allow digits and underscore in compile_uri_template
Allow URI template fields to contain numbers so long as the field name
does not start with a numeral. Example: "test123" is valid, while
"123test" is not.

This patch adds support for this in the legacy template helpers, and also
adds an additional test for these chars for the current router spec.

Co-Authored-By: Kurt Griffiths <mail@kgriffs.com>