Add option to block non-https requests.
Added enforce_https flag to API. If true, middleware is added that
checks whether https is the url scheme, the value of
X-Forwarded-Proto header, or the first proto value of Forwarded header.
If it is not, an exception is raised that can be caught be a user
provided error handler.