fix(testing): Restore httpnow
falcon.testing.httpnow was removed by a previous commit. Restore it as an
alias for the sake of backwards compatibility, placing the actual code
in a new function, falcon.util.http_now().

Also update the docs to refer to the utils directly under falcon.*, rather
than falcon.util.* to promote the fact that developers can access them
in that way.