fix(Response): Instruct browser to remove cookies
Modify the Response object to actually set the expires token for the
cookie that is being unset to a point in the past, causing the browser
to immediately remove the now-expired cookie, which will remove the
cookie from future Request objects.

The previous behavior of Response's unset_cookie function was just to
delete the given name from the SimpleCookie object if it existed.
However, this causes a problem in that the browser is not told to
remove the cookie on its side, causing all future Request objects to
still contain the cookie that was supposed to have been unset/removed.

Co-Authored-By: Kurt Griffiths <mail@kgriffs.com>