feat(api.testing): Testing infrastructure for Falcon Middleware
Added a new testing object - `MiddlewareMock` - for developers to use
for testing their implementations of Falcon Middleware objects. The
`MiddlewareMock` allows a developer to unit test the specific parts
of the middleware in a uniform, reliable manner without having to
rebuild all the required infrastructure to do for every test. Further
it maintains the built infrastructure to be able to allow a developer
to test the stages in successive order.

As part of this work, an instance of the `MiddlewareMock` object is
added to the `falcon.testing.base.TestBase` class so it is easily
available, and clearly distinguishable from the `StartResponseMock`