fix(hooks): Improve detection of requests arg
In order to maintain backwards-compatibility with old hooks that do
not accept a requests argument, the framework performs introspection
when attaching hooks and inserts shims as needed. The introspection
raises an error for hooks that have been defined via
functools.partial().

Fix this by adding support for partial methods and generally making
the introspection more explicit. Modify tests to cover all the different
cases.