fix(Request): Improve access_route perf and error handling
Tune performance of the Request.access_route property, while also
making error handling more consistent, in the case that a header value
is malformed. Also, add a test to ensure that wsgiref passes
REMOTE_ADDR in the environment dict, as has been observed in production
WSGI servers. This is necessary to ensure that Request.access_route's
fallback to REMOTE_ADDR makes sense.