fix(API): Improve and optimize execution of middleware methods
Rework the execution of middleware, responders, and related error
handling in API. This was necessary in order to fix a couple of
bugs (raising errors from resource methods doesn't change status for
middleware, middleware processing isn't always continued when an
instance of HTTPError or HTTPStatus is raised). Along the way,
minor optimizations were made to reduce CPU time in this portion
of the code (offsets additional try..except setup overhead).

Fixes #635