fix(Response): Convert header values to str if needed (#929)
This issue was partially addressed by #490. In this patch, also
ensure that header property setters are converting to str. Also
optimize the logic by inlining it and using str() instead of
str.encode().

Also update docstrings to reference "US-ASCII" instead of
"ISO-8559-1". The latter is technically what is specified in the
WSGI spec, but the HTTP RFCs restrict header text to "US-ASCII" so
there is no reason to expect/allow additional characters.

Fixes #413