fix: check for partials in traverse with verbose option
In order to make the CompiledRouter pickleable, closures were converted
to partials, which introduced a bug in the traverse method with
verbose=True. A check is made for partials and the relevant test case
has been updated to include a regular function along with the partial
generated by the default on_option method.

Tested with Python 2.7 and 3.6.