refactor: Make the CompiledRouter pickleable (#1077)
The CompiledRouter is not pickle-able because the default responder uses
function within functions. As a result, the CompiledRouter cannot be
shared between processes using IPC mechanisms, which utilize pickles
internally. Replace the function within functions with partials of
module level functions so that the CompiledRouter can be pickled. Also,
change the traverse method to use the function wrapped by the partial to
preserve existing functionality.