fix(API): return 500 response to server for uncaught exceptions (#1527)
* fix(API): return 500 response to server for uncaught exceptions

Per newly-understood WSGI spec, applications should trap
internal errors instead of passing them on to the server.
Add a default exception handler for the base Python
``Exception``, which is placed first in the error handler
list (and is thus evaluated last, per LIFO operation).

The 500 error response sends no specific error information
back to the client, but it does log a traceback to
``wsgi.errors``. This logging behavior currently has no tests.

Update error-handling documentation to reflect new default
behavior. No "recipe" is included as yet -- need more details
on requirements here.

Currently, errors are serialized in JSON or XML, depending on
the Accept header in the request. It defaults to JSON if the
request does not prefer XML, even if the app's default media
type is XML. Changing this behavior appears to be non-trivial,
so it is not attempted here, though it was discussed during
development.

* Document interaction between custom error serializer and overriding
error handler for HTTPError

* Add example of adding error handlers

in order of increasing specificity

* Add warning about overriding default ``Exception`` handler

* Correct minor formatting within add_error_handler docstring

* doc(errors.rst): Turn uncaught exception note into a "note"

* doc(faq.rst): Document new default 500 error handling