feat(App): on exception, select most specific error handler available (#1603)
* doc(_handle_error): update NOTE missed in #1527

`_handle_error` returns false when the exception does not match a
custom handler and is not a subclass of ``HTTPError``, ``HTTPStatus``,
or ``Exception``. The update to include ``Exception`` was missed in
PR #1527, which added a default handler for ``Exception``

* feat(API): on exception, select most specific error handler available

Addresses #1514

Rather than selecting error handlers in LIFO order of registration,
select the error handler corresponding to the nearest direct
ancestor of the exception type raised. So, for example, if an app
only adds a custom error handler for the Python ``Exception`` class,
and an ``HTTPForbidden`` error is raised, then we use the
default handler for ``HTTPError`` rather than the more general
``Exception`` handler (which is the pre-existing behavior).

This is implemented by storing error handlers on the API object
as a dict rather than a list and looking them up using the method
resolution order attribute (`__mro__`) on the raised exception
class.

NOTE: This commit only includes the actual implementation and does
not address testing or documentation. I am seeking implementation
feedback before completing those additional changes.

BREAKING CHANGE: Registration of a new error handler for type E
will no longer override previously-registered error handlers for
subclasses of type E. Registration order will no longer matter
*except* when multiple error handlers are registered for the
exact same exception type, in which case the most recently
registered error handler overrides the previous ones.

* test: add test for new error-handling order

* doc(App): Document new error handler precedence logic

* doc(add_error_handler): add versionchanged directive

For the change in error-handler-match strategy

* doc(towncrier): add feature news fragment for error handler precedence