doc/refactor(asgi): Start fleshing out docstrings, tweak doc refs, remove add_error_handler shimming
This patch is a little bit of a grab bag of items that I noticed and
fixed while starting work on fleshing out the docs for the new ASGI
implementation.

Note that there is still quite a bit of work to do in order to integrate
ASGI into the docs. Many docs need to be combed and updated to work with
ASGI and WSGI side-by-side, and there are still plenty of missing docs for
the new ASGI classes and functions.