fix(asgi): Errors caused by not being able to detect cythonized coroutine functions
This patch provides workarounds for the fact that cythonized coroutine functions
can not be detected.

In several places in the framework, we rely on being able to introspect a passed
callable in order to determine how to wrap it or call it later. As a workaround,
this patch gives the caller a way to explicitly indicate that the framework may
assume a coroutine function.

Also, we relax our type checking so that we don't raise an error on Cython
functions; the app developer will simply have to be trusted to pass the correct
type of function if they are using Cython.