perf(asgi): optimize request header decoding
Avoid decoding of header names when possible, and delay value decoding until
it is necessary.

This should reduce the overhead for constructing Request
objects without significantly increasing overall application performance in
the common case, since we reason that applications do not typically access
all request headers that may be present.

The worst case scenario would be an app that accesses the same header more than
a couple of times, in which case we would have to decode said header multiple
times. If this turns out to be a common use case, we will need to look into
memoization or some other mitigation.