docs(examples/asgilook): update aioredis version to 2.0 (#1987)
* update asgilook example + tutorial

Upgrade aioredis library to 2.0

Co-authored-by: Mihai Todor <todormihai@gmail>

* add: asgi process_startup (async) example

Adds a simple example of using ASGI `process_startup`. (This process is now asynchronous.) As per the discussion above, the example pings the server in `process_startup` and closes the pool in `process_shutdown`. Mihai was able to test it by following the tutorial's test app startup process.

Co-authored-by: Mihai Todor <todormihai@gmail.com>

* chore(asgilook): resurrect an accidentally removed newline in a docs snippet

* docs(ASGI tutorial): unpin aioredis in an instruction line

Co-authored-by: Mihai Todor <todormihai@gmail>
Co-authored-by: Vytautas Liuolia <vytautas.liuolia@gmail.com>
Co-authored-by: Mihai Todor <todormihai@gmail.com>