fix(request): prevent IndexError from broken Forwarded header
Fix a wrongful assumption where forwarded_scheme and forwarded_host assume that if a Forwarded header is present, it is valid.
When a Forwarded header cannot be parsed properly (or is plain empty), an empty list is returned.
Instead of an IndexError in forwarded_scheme and forwarded_host, use the existing fallback logic.