fix(tests): spawn servers via sys.executable (#2047)
* fix(tests): spawn servers via sys.executable

Spawn the servers (gunicorn, uvicorn, etc.) during tests through
sys.executable rather than directly, in order to ensure that the test
server is run using the same Python executable as the test suite.
Otherwise, the test servers are started using the Python executable
specified in their shebang and can therefore escape the virtualenv they
are being run in.

This is particularly problem for tests being run from Gentoo ebuilds.
We create a --system-site-packages venv to install falcon in, in order
to test whether it works correctly against the dependencies installed
on the system prior to installing it.  However, the installed packages
such as gunicorn specify /usr/bin/python* shebangs that override
the venv and therefore make these servers unable to find venv-installed
falcon.  Running them via sys.executable ensures that they are started
through Python installed in the venv.