Add boundary types for before hooks
This adds typing information for the before hook as a proof of concept. I used
ParamSpec as a means to introspect the hooks arguments to type check the *args
and **kwargs arguments.

A problem with using ParamSpec is that additional arguments to the before
decorator become impossible (so far as I can see) to specify a type. is_async,
in needs to come after the positional args for backwards compat, but that is not allowed with
ParamSpec.

For now, I've just silenced the type error and we can see if there is a better way