docs(ws-tutorial): add a WebSocket tutorial (#2245)
* docs(ws-tutorial): add tutorial on how to use websockets

Add a tutorial on how to work with websockets in a Falcon app. This is to be read as a quickstart, linking to the reference page for more info.

#2240

* docs(ws-tutorial): move example code to includable files

Include the final code instead of hard-coding it in the docs.

#2240

* docs(ws-tutorial): set max width to around 80

The max width is now set to 80. Not all lines fit, but most do.

#2240

* docs(ws-tutorial): change authentication method for websockets

Authentication is now done via the first-message method instead of headers.

#2240

* docs(ws-tutorial): typo fix

Typo fix.

#2240

* docs(ws-tutorial): Do not generate coverage for the run command

Do not generate coverage for the run command as that's only used when locally running the code.

#2240

* docs(ws-tutorial): add explanation

Add an explanation why the code isn't tested.

#2240

* docs(ws-tutorial): add tests for the example code

Add tests for the example code.

#2240

* chore(wslook): shorten comment lines in `client.py`

* style(wslook): shorten lines in `reports_client.py`

* docs(tutorial-ws): sand some rough edges, fix CI

* docs(WS): add more interlinks to the tutorial

* chore: run tox -e wslook in CI

* docs(ws-tutorial): add coveragerc to the repo

---------

Co-authored-by: Vytautas Liuolia <vytautas.liuolia@gmail.com>