feat (status_codes): update HTTP status constants wrt RFC 9110 (#2330)
* feat (status_codes) update HTTP status code constants wrt RFC 9110
refactor(response): replace falcon.HTTPPayloadTooLarge with falcon.HTTPContentTooLarge

Replace the usage of falcon.HTTPPayloadTooLarge with falcon.HTTPContentTooLarge in the codebase. This change aligns with the recent renaming of the class in the Falcon library. The new class name better reflects the purpose of the error, which is to indicate that the content of the request is too large for the server to process.

Closes #2322

* docs(_newsfragments) documented class name change from HTTPPayloadTooLarge to HTTPContentTooLarge

* fix(status_codes): bring back removed constant for compatibility, fix ruff

* chore: misc touchups

* refactor: add a compat alias HTTPPayloadTooLarge ==> HTTPContentTooLarge

* docs: tweaks newsfragment for HTTP status code updates

---------

Co-authored-by: Vytautas Liuolia <vytautas.liuolia@gmail.com>